backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
gunicorn==23.0.0
hiredis==3.0.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
more-itertools==10.4.0
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pytest==8.3.2
pytest-mock==3.14.0
python-dotenv==1.0.1
python-ulid==1.1.0
redis==5.0.7
requests==2.32.3
scapy==2.5.0
setuptools==72.2.0
SQLAlchemy==2.0.32
types-cffi==1.16.0.20240331
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20240806
types-setuptools==71.1.0.20240813
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
